feat: move isValidEthAddress to lib and add tests for it
