+++ apps/web/src/components/Composer/Actions/CollectSettings/SplitConfig.tsx
@@ -6,10 +6,10 @@ import {
   UsersIcon,
   XCircleIcon
 } from '@heroicons/react/outline';
-import isValidEthAddress from '@lib/isValidEthAddress';
 import { t, Trans } from '@lingui/macro';
 import { HANDLE_SUFFIX, LENSPROTOCOL_HANDLE } from 'data/constants';
 import { useProfileLazyQuery } from 'lens';
+import isValidEthAddress from 'lib/isValidEthAddress';
 import splitNumber from 'lib/splitNumber';
 import type { FC } from 'react';
 import { useAppStore } from 'src/store/app';
